Create Canvas class to allow dynamic render areas
This commit creates the `Canvas` class to allow for dynamic
render areas. This is going to be used in the future for
algorithm visualizations.

The goal here is to make it easy to draw scenes with their own
coordinate systems inside the RenderWindow, at the same time
making it easy to create UI menus without causing interference
with rendering or updating.